Fix button arrangement when using minimap addon
When this addon is installed with a minimap replacement addon, the shape
that the buttons align to is overridden. This happens if minimap.lua is
loaded for any reason, even if improvements are disabled. We solve this
by ensuring that we're only defining GetMinimapShape if it's not yet
been defined by another addon. By checking like this, we don't need to
check for every possible minimap addon by name.